klangscribe-compute/pyproject.toml
[project]
name = "klangscribe-compute"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.58",
    "demucs>=4.0.1",
    "dotenv>=0.9.9",
    "ffmpeg-python>=0.2.0",
    "polars>=1.38.1",
    "torchaudio>=2.10.0",
    "torchcodec>=0.10.0",
    "tqdm>=4.67.3",
]
